requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via starlette
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
idna==3.10
    # via anyio
mysql-connector-python==9.4.0
    # via -r requirements.in
pydantic==2.11.7
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via -r requirements.in
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.35.0
    # via -r requirements.in
